DefaultLegacyOpener: avoid display postprocessing
We want to run the IJ2 open command with all preprocessors (and the
capactiy to include postprocessors) except DisplayPostprocessors, as
these will cause the IJ2 GUI to display an image.. which will then
typically be displayed again by IJ1.

To work around this, we blacklist DisplayPostprocessors by removing them
from the PluginIndex while running OpenFile. The postprocessors are then
returned to the index after execution.

When ready, the LegacyUI should handle this more appropriately.