Move towards standardized contextuality
The various creators, translators and harmonizers are in different
states of being Contextual or not. Often, an unneeded service was being
passed in a constructor just to get its context. This patch makes things
that are Contextual take a Context in their constructor. Things that are
not Contextual just take the service they need.