LegacyJavaRunner: make it work
Unfortunately, we cannot simply hand off to the LegacyService to run
the legacy command because the class is most likely not known to the
PluginClassLoader of ImageJ 1.x.

For now, let's punt and run the class directly (as we used to do in
Fiji's Refresh_Javas plugin). To allow compiling & running classes
that might be known to the PluginClassLoader, we will most likely
jump through hoops. That's next.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>