Create LegacyUI
Create a new ImageJ2 UserInterface implementation: the LegacyUI.

This UI acts as a bridge between ImageJ2 and ImageJ1 - it will only be
active when there is an active ImageJ instance.

The intent of this UI is to handle the display of ImageJ2 command output
via ImageJ1, and to facilitate ImageJ2 commands operating on ImageJ1
structures (such as outputting ImagePluses).