LegacyStatusBar: don't double process events
The LegacyStatusBar now takes advantage of the DefaultLegacyService's
API for setting and checking if a given thread is already processing
status events. If so, we know that we have already entered the given
method of the status bar on that thread, so it is the same event being
processed and we should not re-process it.

Doing so will cause endless numbers of threads to be spawned.