IJ1Helper: also need to trim the leaf command
When adding the final leaf item to our menu structure, it was not being
trimmed of whitespace. This caused it to not match the name that was
registered as an IJ1 command and thus have no associated operation when
selected from the menu (and also causing 2 entries in the IJ1 plugin
list - one with whitespace, one without).

By trimming the actual command name, the menu item now matches the
command and executes properly when selected from the IJ1 menu.