Move some code from the LegacyInitializer to the IJ1Helper
The current strategy for the legacy mode is to let fiji.Main create
a LegacyEnvironment that patches an encapsulated ImageJ 1.x-only
ClassLoader which in turn creates a PluginClassLoader in which the
LegacyInitializer is called to spin up the ImageJ2 context.

In that setting, it does not matter whether the IJ1Helper (initialized
by the DefaultLegacyService) or whether the LegacyInitializer calls the
code to initialize context class loader, icon and application title.

However, our plan is to let the DefaultLegacyService spin up ImageJ
1.x; In that scenario, the LegacyInitializer is no longer called,
but the IJ1Helper is.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>