LegacyImageMap: do not create a new ImageWindow
First, it is wrong to always create just an "ImageWindow". If we have a
hyperstack then we should be creating a "StackWindow".

Second, there are other things that happen in ImagePlus.show, for
example, regarding Window creation. It's possible that we want those
things to happen (e.g. when using the LegacyUI).

Third, creating this window actually prevents consumers (e.g.
LegacyImageDisplayViewer) from calling show(), because the ImagePlus
thinks it's visible already.

So let's just register the given display, and let consuming code decide
what needs to be done with the new ImagePlus.