Avoid depending on EssentialLegacyHooks
The DefaultLegacyHooks need to be loaded in order to execute
DefaultLegacyService's initialize() method, but that is exactly where we
want to call the ij1-patcher that wants to patch the
EssentialLegacyHooks.

So far, that was not a problem because we always had a different class
loader to patch, but we are about to switch to a startup process where
we force ImageJ 1.x to use the same class loader for its plugins as for
its core classes. That means we have to be careful *not* to use the
EssentialLegacyHooks class prematurely.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>