This is the big switch to the new startup procedure
The intent of this switch is to make Fiji start up ImageJ2 in the future
instead of starting ImageJ 1.x and then letting the LegacyInitializer
spin up a corresponding ImageJ2 context.

This implies that we will have to reuse the same class loader for
ImageJ2, ImageJ 1.x and ImageJ 1.x' plugins. As a consequence, we will
have to discontinue support for Help>Refresh Menus, but given that
there are now much better ways to develop plugins in Integrated
Development Environments (such as Eclipse, Netbeans, etc), that is
not too big of a concession.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>