Exclude the script templates from getting license headers
The script templates are intended to be as accessible as possible. As
a consequence, they are *not* to have a license header.

For the record: the templates as of Fiji's script editor were always
intended to be freely usable. In other words, they were licensed
implicitly under the CC Zero license. For details about that license,
see http://creativecommons.org/publicdomain/zero/1.0/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>