Help diagnose multiple contexts
For the moment, there can only be one LegacyService per class loader.
Let's make sure that we're giving enough information to diagnose what
went wrong when a second legacy services was about to be instantiated
nevertheless.

While at it, make sure that resetting the current instance is guarded
in a synchronized block just like setting it is.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>