Do not prevent ImageJ 1.x from quitting
Since we switched to the LegacyUI, we need to let the ImageJ context
know that it should be disposed. To prevent ImageJ 1.x from trying to
run the quit procedure *twice* (which *almost* works, despite failing
to check that it already quit, resulting in a nasty NullPointerException
when trying to save the preferences and calling a method on the return
value of IJ.getInstance() *without* verifying first that it is not
null), we force ImageJ 1.x to quit only once.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>