IJ1Helper: avoid duplicating recent list
If the item was the most recently used (at index 0), it was added again,
bumping the last item off the list.

Instead, we want to do nothing if it's already the most recently used,
move the item if it's on the list, and replace the oldest item if the
new item was not on the list.