Return path if selected
If, during the DefaultLeagcyOpener#open execution, we get to the point
where the user selects a path, we will now return that path if we fail
to open it. This will give IJ1 a chance to open the selected path,
instead of forcing them to select a new path with no message indicated.