Open failed paths via IJ 1.x
If we get a String back from our LegacyOpener, we will now attempt to
open it directly via IJ 1.x. Ideally we would return a value that makes
sense or update an argument, but this is not possible for the "open"
command due to the case logic in the ij Commands class.