Don't overwrite existing dimensions
The GrayImagePlusCreator, if given say an "XYZ" image, would say "I
found dimensions 0, 1 and 3" referring to the indices in the "XYCZT"
order. However, it would store these in indices 0,1,2 of a 5-d axis
array, and the min/max values would be at indices 0,1,3.

The index 3 of the min/max array corresponds to a -1 value in the axis
array, indicating the axis was not found. Thus causing the Z axis to be
disregarded when padding the dimensions to XYCZT.

The correct behavior is to determine if, for each axis index, determine
if that axis was found in the original ImgPlus, and only pad if it was
  not found.

Closes #36