Disable the OptionsSynchronizer completely
It takes a very long time to synchronize the options wholesale, and while
it was well-intentioned to synchronize the options with the legacy mode
at all, it is highly dubitable that we actually need to finagle the
modern options into the legacy options.

Better to keep them separate, and at the same time save time (also
maintenance time because we no longer have to maintain that beast).

As a bonus, ImageJ2 does not inherit compatibility warts such as the
problematic "black is foreground" setting that cause so many macros to
fail miraculously.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>