Make ImageJ2 options non-modal
We are about to add a button to let the user ask to be shown the welcome
message again.

Alas, if the ImageJ2 options dialog is modal, the user cannot even
scroll, let alone put the welcome message into the foreground.

Better to have the options dialog be non-modal; that also lets people
switch off the SCIFIO option just to load a single file that SCIFIO might
not be able to load yet and immediately switch SCIFIO back on.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>