Refactor color image support
We have a method in LegacyUtils called "isColorCompatible".. so we
should use that to determine if a dataset should be made with a
ColorImagePlusCreator or Grayscale.

Also, we were ALWAYS making RGB datasets, and if there was more than one
slice we would try to composite them for some reason. The correct
behavior is, if we have 1-3 channels and an RGB dataset, make an RGB
ImagePlus. If we have 2-7 composite channels, we can make a
CompositeImage in composite mode.

The ColorImagePlus creation pathway has been updated accordingly to
match these new specifications.