IJ1Helper: close non-image windows
In the ImageJ 1.x close process, non-Image non-Frame windows are not
closed. We need to ensure they are closed completely to properly shut
ImageJ2 down. If we do not, non-Frame non-Image dialogs like the
ContrastAdjuster will never close (in which case, the thread it spawns
will never end, and Fiji will not shut down).