Refactor PixelHarmonizers for extra dims
Now that unknown dimensions are stored in the Time axis we need our
pixel harmonizers to properly read the unknown dimensions from the
source image.

This fixes the Gray and ColorPixelHarmonizers so that they compute the
position of each T index in the potentially multi-dimensional space of
the source image, and query the appropriate position as needed.