Initialize legacy-compatible commands lazily
We really only need this when the legacy UI initializes. This means we
can discover the legacy-compatible commands lazily.

In the next commits, we will add support for legacy-compatible scripts;
at least for the ImageJ 1.x macro language, we need that lazy
initialization anyway, otherwise the script service (called from the
legacy service) will try to inject the context into the ImageJ 1.x macro
engine (which in turn needs a *fully initialized* legacy service).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>