Refactor how legacy-compatible command are handled
We are about to introduce support for legacy-compatible scripts; the
corresponding ScriptInfos will be handled very similarly to the way
CommandInfos are handled. Let's refactor that code in preparation.

In particular, do not require a list of CommandInfo instances, but
instead a map from keys to ModuleInfo instances. The keys will put into
ImageJ 1.x' commands table by which we identify the infos to execute
later. By using a sensible convention ("Command:<class-name>"), this
adds incidental information to the command launcher.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>