IJ1Helper#runMacro: ensure that Macro.getOptions() works
ImageJ 1.x has this funny homebrewn way to provide thread-local
macro options (without using the ThreadLocal class at all). It
also poses a completely unnecessary restriction on the the thread
name. To support that in the broader context, we have to make
sure that the current thread name conforms to that requirement.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>