Safeguard against infinite loops in getThreadAncestors()
When we try to help ImageJ 1.x read its macro options, we traverse the
parent threads until we find one with macro options, or until the chain of
parent threads is broken.

There are subtle ways, however, to have infinite loops there.

Instead of looping indefinitely, just stop when we had a look at all
available threads in the parent chain.

This problem could be triggered reliably by switching to modern mode and
then hitting the "o" key to open a file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>