DefaultLegacyHooks: make closeAllWindows do more
When ij.WindowManager#closeAllWindows() is called, there are some
windows that it fails to close down. In particular, any non-Frame
non-image windows such as PlugInDialogs are not handled. For example,
the ContrastAdjuster spawns a new thread which will only terminate
when the ContrastAdjuster window is disposed, but the
WindowManager#closeAllWindows() method will not do that.

Further, ImageJ1 does not grant the user the opportunity to confirm
dismissal of certain types of windows. Specifically: under certain
circumstances, any non-image window which is not an
ij.plugin.frame.PlugInFrame, ij.plugin.frame.Editor or
ij.text.TextWindow will be forcibly dismissed via its dispose() method.

See:
https://github.com/imagej/ImageJA/blob/v1.49c/src/main/java/ij/WindowManager.java#L391-L401

As such, and keeping in mind that ImageJ1 is intended to run as an
end user application in its own dedicated JVM, this commit makes the
closeAllWindows routine go the whole nine yards, and really close down
_all_ Windows in the JVM, while _always_ giving the user the opportunity
to cancel closing them down. Of course, it would be ideal to close down
only those windows associated with ImageJ1, but unfortunately we have no
way of knowing for certain which windows those are, since an ImageJ1
plugin can create and show any arbitrary Windows of its choosing,
without registering them with the WindowManager.