When shutting down IJ1, dispose the context too
There are essentially two ways ImageJ1 can be shut down:

1) From the ImageJ1 side; i.e.: ij.ImageJ#quit()
2) From the SciJava side; e.g.: LegacyService#dispose()

When IJ1 is shut down via its quit() routine, it should be done
"softly", asking the user to confirm the closure of various windows,
etc., before finally disposing of ImageJ1's resources as well as the SJC
context. System.exit should be called under exactly the same
circumstances as when IJ1 would normally do it.

When IJ1 is shut down via the legacy service, it should be done "hard",
without prompting the user to confirm anything. This form of shutdown
should not trigger a disposal of the SJC context, since A) it might
already be in the process of being disposed; and B) if it isn't already,
then it is almost certainly not desired for it to happen.

This change makes sure that when IJ1 is shutting down in the former
case, the SJC context gets disposed too, but that when IJ1 is shutting
down in the latter case, an SJC disposal is _not_ triggered.