Make imagej-ui-swing non-optional
The original thinking (which we unfortunately did not write down
anywhere that we could find) was probably that depending on
imagej-legacy should not require the project to inherit the
imagej-ui-swing dependency, since it is a UI-specific component.

However, the relevant legacy code using the net.imagej.ui.swing
classes (in particular, the TextEditor of the Script Editor) was not
careful to catch all ClassNotFoundExceptions etc. in cases where the
imagej-ui-swing was not also present on the classpath.

Further, the imagej-legacy project's purpose is to manage the
interaction with ImageJ1, which is inherently an AWT & Swing based
application. So there is no advantage to restricting dependence on
Swing-based libraries anyway.

Better is to simply declare it as a normal dependency. Then if someone
wants to exclude the dependency, they can do so... at their own risk.

This change eliminates the nasty exception stack traces from the output
of the ij1-quit integration test.