Skip shutdown regression tests in headless mode
In headless mode, there is really very little we can test: there is no
main window (even hidden) to begin with, we cannot properly dispose
anything, and in headless mode the shutdown procedure is completely
different to begin with.

So let's just skip the regression tests in headless mode, relying on the
Windows-specific Jenkins job to test the shutdown hooks (the Windows node
does not run headlessly).

Thanks, Jizelle Jenkins, for pointing out the problem!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>