Let ImageJ2 quit
It is true that Fiji won't quit. But that refers to the project, not the
application.

We worked pretty hard on fixing the problem where the application would
not quit, but we forgot one crucial thing: when the user quits the
application, there is always one window that is currently not disposable:
the ImageJ 1.x window itself!

	:-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>