Update window closing to use CloseConfirmable
Dispatching a WINDOW_CLOSING event can lead to deadlocks due to the
closeAllWindows method being synchronized. For example, if a listener
tries to remove itself from the WindowManager (also synchronized) the
EDT will be blocked until closeAllWindows completes, so calling dispose
as part of the interception will always deadlock at that point.

CloseConfirmable is an opt-in close cleanup mechanism. Any windows
implementing this interface will have their confirmClose method called,
and the close operation will continue based on the return value of that
method.