ImageJ2Options: use a better name for I/O option
Firstly, the option is not limited to SCIFIO. It actually toggles the
new-style ImageJ2 I/O, which includes but is not limited to SCIFIO.

Further, writing "isUse" rankles me. Yes, it is best practice for
boolean accessors to begin with "is", but that assumes the variable is
named in a nounish way rather than a verbish way. To do it 100%
correctly means choosing a variable name to which "is" can be prepended
sensibly (e.g., "scifioEnabled").