ImageJ2Options: remove unused & dangerous constant
String constants get inlined, which means that once a public one is
declared, for backwards compatibility, it should never be changed. But
we don't really want people using this constant to access the options
field by name anyway -- we have an accessor method for that. So let's
get rid of the constant before it does too much damage downstream.