Fix ImagePlus/Display synchronization
We always want to create a shadowing Display for a given ImagePlus, but
we are also shadowing Displays with ImagePluses. Care needs to be taken
so that the appropriate shadowing structures are created and disposed at
appropriate times.

This commit adds safeguards for construction and deletion of Displays
and ImagePluses to avoid potential infinite loops or NPEs due to
recursive instantiation or multiple deletions.