Broaden the scope of "legacy mode"
The isLegacyMode method will now return true if there is a non-null
ImageJ 1.x instance, instead of worrying about whether it was visible or
not.

As it was, if we were running headless then isLegacyMode would always
return false - which is not the behavior we wanted. This may be a bit
aggressive, but it is more correct and in line with our "we don't care
about modern mode" attitude.