Fix premature initialization of ImageJ App
The net.imagej.app.ToplevelImageJApp was being forced to initialize
during the DefaultLegacyService's initialize() process, which caused it
to have a null LegacyService injected, which caused it to report only
the ImageJ2 version, and not the combined IJ2/IJ1 version.

The fix is to avoid leaning on services as much as possible during
DefaultLegacyService#initialize().