Be very sure ImagePluses are removed
It is not enough to rely on ImageWindow#dispose... you HAVE to set the
active image as null in the WindowManager.

However, using the deleteImageMethod may be too extreme as it leads to
double ImagePlus closing. Thus, if we don't want to delete the
ImagePlus, we will still remove its ImageWindow from the WindowManager.

Fixes http://fiji.sc/bugzilla/show_bug.cgi?id=819