LegacyPlugInInfo: only add arg to known plugins
In many cases, the arg is actually a file path or similar actual
"argument" to the legacy plugin. The common exception is the
ij.plugin.Commands class, which actually accepts the subcommand to
execute as its arg. We want to preserve such subcommands, but not, in
general, arguments to plugins like file paths.