Do not let the usage statistics prevent bare .class plugins from running
There is a "cute" logic in ImageJ 1.x allowing to run bare .class plugins
even if they are in subdirectories of the plugins/ directory. This
requires some hackery to work, thanks to the ImageJ 1.x patcher. But as
long as that hackery has not been run, the class cannot be run properly.

For the moment, simply ignore such plugin classes when gathering usage
statistics.

This should fix http://fiji.sc/bugzilla/show_bug.cgi?id=891 reported by
Aryeh Weiss.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>