Add an option to toggle IJ1<->IJ2 sync
This option is off by default, to avoid disrupting IJ1-based workflows.
When the option is enabled, newly created ImagePlus objects will be
synced to ImageJ2 ImageDisplays. Ideally, this would be a lazy/cheap
operation, but at the moment it is too expensive.

Fixes Fiji bug 899:
http://fiji.sc/bugzilla/show_bug.cgi?id=899

Or at least: it avoids the problem for users by default, which is good
enough for now.