ImageJ2Options: add missing callback
This fixes a problem where toggling the sync state does not actually do
anything because run is not called in response. The callback is only
needed because the ImageJ2Options dialog is Interactive, meaning it has
no OK or Cancel buttons. (And it is only Interactive so that it does not
block the application when the "Welcome to ImageJ2" window appears.)