DefaultLegacyOpener: use registerDataset method
This was being done incorrectly - creating an ImagePlus from a Dataset,
then registering it with the LegacyImageMap, which created a new
Display, which created a new Dataset from the (shadowing) ImagePlus.

We can now simply go from Dataset to ImagePlus and let the
LegacyImageMap take care of creating (but not showing) a display, and
mapping appropriately.