Revert "IJ1Helper: always redirect stdout and stderr"
This reverts commit 2447ea0809fefefda8c47c423e703a4223fd1078.

This address Fiji bug 984:
http://fiji.sc/bugzilla/show_bug.cgi?id=984

As well as several complaints on the mailing lists.

It seems that many people use the ImageJ1 Log window for macro output,
and conflating anything else into that window disrupts workflows.

So, we will need to find another way to preserve vital information
dumped to stdout and stderr. Perhaps a dedicated Console window? It
should probably not appear automatically though unless an exception
is detected. See issue #97 for details.