Fix ImagePlus.show deadlocks
Using the DisplayService was causing ImagePlus.show to be called from
the EDT. If this happens in a macro, it will cause a 2 second timeout
deadlock to be hit for every call to show.

As a short-term workaround we can directly call ImagePlus.show,
bypassing the DisplayService. This unfortunately also bypasses any
extension points for viewers, displays, etc...

This commit should be revertable when
https://github.com/scijava/scijava-common/issues/144 is resolved.