Use full file path if no dot
The LegacyUI.chooseFile method erroneously assumed a dot (.) would be
present, causing ArrayIndexOutOfBounds exceptions if one was not found.
As this is just the default string to populate the file chooser, there
is no need to require a dot in the base path name.

Instead, if no dot is present we will just use the full path.