ImageJ2Options: purge horrible syncEnabled option
Rather than making it a persistent setting, which to this day users
still somehow have enabled by accident -- or perhaps in some cases, on
purpose without understanding the implications, despite the warnings --
let's make the setting toggled by a system property which must be very
intentionally set via the command line or by a script.

The hope is that this change will finally squelch the continuing bug
reports on the ImageJ mailing list and in the Fiji BugZilla, related to
this regrettably neglected and still-broken feature.