ImageJ2Options: reduce intensity of warnings
Now that the "Enable ImageJ2 data structures" option has been removed,
it is less pivotal to discourage people from enabling the other
remaining option: SCIFIO. While SCIFIO does fail with some datasets,
it also enables support for others. So it is probably sufficient to:

A) Have SCIFIO turned off by default; and
B) Mark the feature as "BETA!" in all caps with a bang.

It is definitely the case that the SCIFIO option _does_ help some users
with certain datasets such as problematic TIFFs. And it is also the case
that when a user enables SCIFIO and then gets an exception and reports
it as a bug, this is extremely helpful, since our goal is for SCIFIO to
someday completely replace the legacy ImageJ1 file I/O behavior.