Make macro parameter accesses in the IJ1Helper non-static
The main idea of the IJ1Helper is to encapsulate ImageJ 1.x accesses so
much that we will eventually be able to separate all such accesses enough
to be able to use two different Contexts with separate LegacyServices in
the very same class loader.

That also means that static accesses are a big nonono.

Technically, this is a backwards-incompatible change, requiring a major
version bump, but we're still in beta.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>