LegacyService: group unused services together
It is unfortunate that we need to declare any service parameters which
never get used. Most services do not need to do this. But the
LegacyService does a whole lot of things in its initialize() method --
things that lean on helper classes which need those services. Hence, the
legacy service requires those particular services to be available.

At some point, we should convert the LegacyService over to lazy
initialization, to avoid this issue. But not now.