Eliminate LegacyService as an interface
The DefaultLegacyService is required all over the codebase, thanks to a
proliferation of public methods declared in DefaultLegacyService only.
As things stand, it would be impossible to actually implement an
alternative LegacyService. So let's stop pretending it's possible, and
instead simply make the LegacyService the one and only such service.

Fixes #53.