Fix the `-batch <file>` handling
Triggered by a tweet of Stephan Preibisch, this developer tried to run a
macro in batch mode via

	.../ImageJ-win32 -batch abc.ijm

and was surprised that it did not work. The reason it did not work was a
completely borked logic behind the `-batch` handling, which assumed that
`.../ImageJ-win32 -eval ... -batch` was the only use case.

This patch fixes that logic to work correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>