Prepare IJ1Helper for the upcoming SingleInstance listener
In ImageJ1, there is an option in Options>Misc...>Run single instance
listener; When checked, it tells ImageJ1 to look for an existing ImageJ
instance when starting up, and passing the command-line options if one
is found.

We will reinstate that functionality in the next few commits and need to
prepare the IJ1Helper for that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>