Add the equivalent of ImageJ1's OtherInstance class
The purpose of this class -- which we now call SingleInstance because it
is more appropriate -- is to start an Remote Method Invocation (RMI)
server when ImageJ is started, listening for commands being sent by
subsequent invocations of the ImageJ executable.

Essentially, it makes ImageJ a singleton, when that "Run single instance
listener" option is checked.

We still need to *use* this class, of course, which is the purpose of
the next commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>