Open images passed via the command line again
Due to recent changes in the `openImage()` method, images are not shown
when loaded as long as the UI is not shown. This is the case when images
are loaded as part of command line parameter parsing.

Let's force images to be shown when we know that the UI will be shown
eventually.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>