Use whitelist for disposing uncofirmable windows
We knew that adding ALL windows to a list of things to dispose could
cause problems. In particular, it causes closing to not play well when
launching Fiji from within other Java applicaitons (particularly
MATLAB).

Instead we will dipose ij.* explicitly, in addition to anything
CloseConfirmable (as always). This may still cause Fiji plugins to
create some issues - but in those cases, the offending classes should be
updated to implement CloseConfirmable.