IJ1Helper: key menu cache on entire path prefix
Rather than keying on only the current menu of interest, we use the
entire menu prefix, to avoid clashes when two menu paths share a common
menu entry.

For example, suppose you have two commands with menu paths:
- One > Stuff > Foo
- Two > Stuff > Bar

Since they both share "Stuff" as part of their menu paths, the previous
logic would create one single "Stuff" menu, cache it, and then reuse it
for the other command -- erroneously.

We really want to be keyed on "One > Stuff" and "Two > Stuff" prefixes.
To be really safe though, we use "\n" instead of ">" since "\n" is
unlikely to appear as part of a menu path (and probably would result in
other issues if it did).

Fixes scijava/scijava-common#221.