SingleInstance: don't include monitor in stub file
Including the monitor ref (e.g. *.0) of the current display in the stub
file name leads to unintuitive behavior: such as two instances popping
up when one is opened via command line and the other via UI.

The intent of this feature is to give the user a single instance
experience, regardless of how ImageJ is launched.