SingleInstance: fix stub file checking
For the SingleInstance listener, if a stub file exists and reading it
produces a valid remote object, then we infer that another instance is
running and pass it all arguments (if any).

If the stub file is invalid or not found, then the running instance is
inferred to be the canonical instance and a RMI server is launched.