IJHelper#recordOption: escape special chars
When recording a key value pair such as:

    file = C:\Users\jdoe\image.tif

We want the macro recorder to record the parameter as:

    "file=C:\\Users\\jdoe\\image.tif"

with escaped backslashes.

Otherwise, we end up with an invalid string literal.

Since ImageJ 1.x's Recorder.recordOption method does not do any
escaping, we do it in the IJ1Helper instead, to ensure safety.

We also escape the double quote symbol (") for similar reasons.

While we're at it, we also replace common special characters
with their corresponding Java string literal escape sequences.
E.g., tabs become \t and newlines become \n.

But we do not (yet ;-) go so far as to escape Unicode.

See also the previous commit, as well as scijava/scijava-common#235.