Update JavaRunner plugin to new SJC CodeRunner API
The scripting-java JavaRunner plugin type moved into SciJava Common
as the more general CodeRunner in the org.scijava.run package.

This commit updates the plugin to use the new API.

It still does not provide complete support for running legacy
ImageJ 1.x plugins via the --run flag, but it gives us the basics.