Allow more than one simultaneous LegacyService
It is extremely annoying that you cannot create a second SciJava Context
when the LegacyService is on the classpath. The rationale for that is
that there is only a singleton instance of ImageJ 1.x, so there should
only be one LegacyService allowed. But as things stand, that restriction
completely forbids additional Contexts from being created, unless great
care is taken to exclude the LegacyService during the context creation
-- and doing that is not easy, since context creation is additive only,
with no facility to blacklist unwanted services.

This change enables multiple simultaneous LegacyService instances to be
created, by making additional LegacyServices behave in a restricted way,
without access to IJ1. In this way, additional Contexts can now be
created, as long as they do not need to perform legacy-specific actions.