LegacyImageJApp: avoid LegacyService @Parameter
Unfortunately, the LegacyService does a lot of work in its initialize()
method, which happens before the Context has finished spinning up.
It ends up triggering a call to the AppService, which loads the App
plugins, which creates the LegacyImageJApp, which tries to inject a
LegacyService before one exists in the Context... so to avoid this
circularity, for now, let's not inject the LegacyService that way.