IJ1MacroEngine: stop handling 'result' specially
In recent versions of SciJava Common, the script framework handles
return values in a special way, assigning them into an output called
'result' when no other outputs are declared.

See:
https://github.com/scijava/scijava-common/compare/1aea390b9bb3b575f7f28bc15c6e3a9f24bdfdaa...1e1020437a9d9a014de23d906f7ed9bbcaf2bb18

The return value is now stored into an output called 'result' when no
outputs are declared. And if the 'result' variable is assigned in the
script, that assigned value will be used, rather than defaulting to the
script's return value.

Hence, there is no need to do anything clever with the 'result'
variable in the IJ1 macro language specifically.