IJ1MacroEngine: improve image ID extraction logic
In the case of Number (which is almost certainly a Double here),
we should simply call intValue(), rather than converting to
String and then parsing an Integer. The String route is not only
less efficient -- it is also unsafe, since Double.toString() for
integer values includes ".0" after the integer value, which does
not parse successfully using Integer.parseInt(s).

And in the case of String, we should catch NumberFormatException
in case the string is not actually a valid integer.

Previously, the engine would crash if an image ID could not be parsed.
This could (theoretically -- I did not test) occur with something like:

    // @OUTPUT ImagePlus result
    result = "hello"

Because "hello" will not parse as an integer.