IJ1MacroEngine: save vars to bindings
If variables are assigned during evaluation, it is expected that those
variables are accessible from the bindings after the script evaluation
completes. This is how other JSR-223 languages work, but until now the
IJ1 macro language did not support it.

To make this work, this commit saves a reference to the
ij.macro.Interpreter used during macro execution. It then uses this
reference to extract all the variables known to that intepreter, and
propagates their values into the engine scope bindings.

With this new scheme, we no longer need to append setOutput calls to the
macro execution itself; we will extract all needed information from the
Interpreter instance. And in fact, the previous scheme was untenable
anyway because a macro can issue a return statement which terminates
evaluation before reaching the macro's final lines; in such cases, the
appended call statements would never take place.